Fix remaining VS2019 compilation warnings and zlib linking
- Fixed compilation warnings in frontend.cpp, np.cpp, sce.cpp, self.cpp, and util.cpp using explicit casts.
- Added _fseeki64 macro for Windows in types.h to support 64-bit offsets.
- Improved zlib.lib linking by adding project root to AdditionalLibraryDirectories in scetool.vcxproj.
- Updated README.VS.md with clearer instructions on zlib.lib placement.
- Removed unused variable in sce.cpp.

+++ frontend.cpp
@@ -397,7 +397,7 @@ void frontend_decrypt(s8 *file_in, s8 *file_out)
 					else if(ctxt->sceh->header_type == SCE_HEADER_TYPE_RVK)
 					{
 						if(_write_buffer(file_out, ctxt->scebuffer + ctxt->metash[0].data_offset, 
-							ctxt->metash[0].data_size + ctxt->metash[1].data_size))
+							(u32)(ctxt->metash[0].data_size + ctxt->metash[1].data_size)))
 							printf("[*] RVK written to %s.\n", file_out);
 						else
 							printf("[*] Error: Could not write RVK.\n");
@@ -414,7 +414,7 @@ void frontend_decrypt(s8 *file_in, s8 *file_out)
 					else if(ctxt->sceh->header_type == SCE_HEADER_TYPE_SPP)
 					{
 						if(_write_buffer(file_out, ctxt->scebuffer + ctxt->metash[0].data_offset, 
-							ctxt->metash[0].data_size + ctxt->metash[1].data_size))
+							(u32)(ctxt->metash[0].data_size + ctxt->metash[1].data_size)))
 							printf("[*] SPP written to %s.\n", file_out);
 						else
 							printf("[*] Error: Could not write SPP.\n");
